+++ src/main/java/com/tms/controller/UserController.java
@@ -13,9 +13,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
+
 import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 
 @RestController
@@ -38,10 +40,10 @@ public class UserController {
     @GetMapping("/{id}")
     public ResponseEntity<User> getUserById(@Parameter(description = "Эта та самая id которую нужно передать") @PathVariable int id) {
         User user = userService.getUserById(id);
-        if (user == null){
+        if (user == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        return new ResponseEntity<>(user,HttpStatus.OK);
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     @Operation(summary = "Создание пользователя")
@@ -51,33 +53,67 @@ public class UserController {
             @ApiResponse(responseCode = "500", description = "Ошибка на сервере.")
     })
     @PostMapping
-    public ResponseEntity<HttpStatus> createUser(@RequestBody User user, BindingResult bindingResult) {
+    public ResponseEntity<?> createUser(@RequestBody @Valid User user, BindingResult bindingResult) {
         if (bindingResult.hasErrors()) {
             for (ObjectError o : bindingResult.getAllErrors()) {
                 log.warn("We have bindingResult error : " + o);
+                return ResponseEntity
+                        .status(HttpStatus.BAD_REQUEST)
+                        .body("ВАЛИДНОСТЬ ДАННЫХ:  " + o);
             }
         }
         userService.createUser(user);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
+    @Operation(summary = "Изменение пользователя")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Все супер!"),
+            @ApiResponse(responseCode = "400", description = "Не получилось достать по id..."),
+            @ApiResponse(responseCode = "500", description = "Ошибка на сервере.")
+    })
     @PutMapping
-    public void updateUser(@RequestBody User user) {
+    public ResponseEntity<?> updateUser(@RequestBody @Valid User user, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            for (ObjectError o : bindingResult.getAllErrors()) {
+                log.warn("We have bindingResult error : " + o);
+                return ResponseEntity
+                        .status(HttpStatus.BAD_REQUEST)
+                        .body("ВАЛИДНОСТЬ ДАННЫХ:  " + o);
+            }
+        }
         userService.updateUser(user);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-
+    @Operation(summary = "Удаление пользователя по id")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Все супер!"),
+            @ApiResponse(responseCode = "400", description = "Не получилось достать по id..."),
+            @ApiResponse(responseCode = "500", description = "Ошибка на сервере.")
+    })
     @DeleteMapping("/{id}")
     public ResponseEntity<HttpStatus> deleteUser(@PathVariable int id) {
-        userService.deleteUser(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        User user = userService.getUserById(id);
+        if (user == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        } else {
+            userService.deleteUser(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-   @GetMapping
+    @Operation(summary = "Просмотр всех пользователей")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Все супер!"),
+            @ApiResponse(responseCode = "400", description = "Не получилось достать по id..."),
+            @ApiResponse(responseCode = "500", description = "Ошибка на сервере.")
+    })
+    @GetMapping
     public ResponseEntity<ArrayList<User>> getAllUser() {
         ArrayList<User> list = userService.getAllUsers();
-       return new ResponseEntity<>(list, (!list.isEmpty()) ? HttpStatus.OK : HttpStatus.NOT_FOUND);
-   }
+        return new ResponseEntity<>(list, (!list.isEmpty()) ? HttpStatus.OK : HttpStatus.NOT_FOUND);
+    }
 
 
 }
